批量清理warning(100+->67): Icons.AutoMirrored迁移(11处)+逻辑冗余删除(35处safe-call/elvis/死分支)

+++ app/src/main/java/top/hsyscn/opedrgent/network/MapTileFetcher.kt
@@ -54,8 +54,8 @@ object MapTileFetcher {
             DebugLog.i("MapTileFetcher: fetching map at $lat, $lon zoom=$zoom ${widthTiles}x${heightTiles} tiles")
 
             val (centerTileX, centerTileY) = latLonToTile(lat, lon, zoom)
-            val offsetX = ((lonToX(lon, zoom) - centerTileX) * TILE_SIZE).toInt().coerceIn(0, TILE_SIZE)
-            val offsetY = ((latToY(lat, zoom) - centerTileY) * TILE_SIZE).toInt().coerceIn(0, TILE_SIZE)
+            val offsetX = ((lonToX(lon, zoom) - centerTileX) * TILE_SIZE).coerceIn(0, TILE_SIZE)
+            val offsetY = ((latToY(lat, zoom) - centerTileY) * TILE_SIZE).coerceIn(0, TILE_SIZE)
 
             val startX = centerTileX - widthTiles / 2
             val startY = centerTileY - heightTiles / 2
